# backend/internal/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Habitus Partner API",
    "version": "2026-09-06",
    "summary": "Подбор жилья по запросу словами — как встраиваемый сервис.",
    "description": "Habitus разбирает запрос на естественном языке, подбирает объекты и объясняет каждый выбор фактами о районе: сколько ребёнку идти до школы, куда выходят окна, что вокруг по вечерам.\n\nPartner API отдаёт то же ядро, что работает в самом продукте: поиск, разбор запроса, досье объекта с проверяемыми утверждениями, гео-слои города, управление собственной витриной и заявки покупателей.\n\n**Главное правило сервиса — не выдумывать факты о городе.** Если замера нет, поле придёт как null, а блок разбора деградирует или исчезнет. Нулей вместо отсутствующих измерений здесь не бывает; модельные оценки всегда помечены как proxy.",
    "termsOfService": "https://habitus.city/terms",
    "contact": {
      "name": "Поддержка интеграций",
      "email": "api@habitus.city",
      "url": "https://habitus.city/docs"
    },
    "license": {
      "name": "Проприетарная лицензия",
      "url": "https://habitus.city/terms"
    }
  },
  "servers": [
    {
      "url": "https://api.habitus.city",
      "description": "Продакшен"
    },
    {
      "url": "http://localhost:8080",
      "description": "Локальная разработка (docker compose up)"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/partner/v1/listings": {
      "get": {
        "operationId": "listListings",
        "tags": [
          "Витрина"
        ],
        "summary": "Объявления партнёра",
        "description": "Ваш каталог в системе. Чужих объявлений здесь не бывает: выборка всегда ограничена вашим аккаунтом.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "publishing",
                "published",
                "unpublished",
                "failed"
              ]
            },
            "description": "Фильтр по статусу."
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница объявлений.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/List"
                }
              }
            }
          },
          "401": {
            "description": "Ключ не передан, не распознан, отозван или истёк.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "invalid_api_key": {
                    "value": {
                      "error": {
                        "type": "authentication_error",
                        "code": "invalid_api_key",
                        "message": "Ключ API не распознан. Проверьте заголовок Authorization: Bearer hab_live_…",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ключу не хватает права, либо доступ партнёра приостановлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "insufficient_scope": {
                    "value": {
                      "error": {
                        "type": "permission_error",
                        "code": "insufficient_scope",
                        "message": "Ключу не хватает права search:read",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Превышен лимит. Заголовок Retry-After говорит, через сколько секунд повторять.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "type": "rate_limit_error",
                        "code": "rate_limit_exceeded",
                        "message": "Превышен лимит запросов (120 в минуту). Повторите через 12 с",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не прошёл проверку. Поле названо в param.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "validation_error": {
                    "value": {
                      "error": {
                        "type": "invalid_request_error",
                        "code": "validation_error",
                        "message": "query обязателен",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "query"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createListing",
        "tags": [
          "Витрина"
        ],
        "summary": "Завести объявление",
        "description": "Создаёт карточку и, если передан publish: true, сразу отдаёт её в поиск. Публикация считает эмбеддинг объявления и занимает секунды.\n\nПровал публикации не отменяет создание: объявление останется со статусом failed и причиной в поле error — уже принятые данные не теряются, повторить можно вызовом /publish.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingInput"
              },
              "examples": {
                "minimal": {
                  "summary": "Минимум полей",
                  "value": {
                    "city": "msk",
                    "coordinates": [
                      37.6156,
                      55.7522
                    ],
                    "price": 17500000,
                    "area": 54.2,
                    "rooms": 2,
                    "level": 5,
                    "levels": 12,
                    "address": "Москва, ул. Пример, 1",
                    "publish": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Объявление создано.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не прошёл проверку. Поле названо в param.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "validation_error": {
                    "value": {
                      "error": {
                        "type": "invalid_request_error",
                        "code": "validation_error",
                        "message": "query обязателен",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "query"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ключ не передан, не распознан, отозван или истёк.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "invalid_api_key": {
                    "value": {
                      "error": {
                        "type": "authentication_error",
                        "code": "invalid_api_key",
                        "message": "Ключ API не распознан. Проверьте заголовок Authorization: Bearer hab_live_…",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ключу не хватает права, либо доступ партнёра приостановлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "insufficient_scope": {
                    "value": {
                      "error": {
                        "type": "permission_error",
                        "code": "insufficient_scope",
                        "message": "Ключу не хватает права search:read",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Превышен лимит. Заголовок Retry-After говорит, через сколько секунд повторять.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "type": "rate_limit_error",
                        "code": "rate_limit_exceeded",
                        "message": "Превышен лимит запросов (120 в минуту). Повторите через 12 с",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/partner/v1/listings/{listing_id}/publish": {
      "post": {
        "operationId": "publishListing",
        "tags": [
          "Витрина"
        ],
        "summary": "Опубликовать",
        "description": "Отдаёт объявление в поиск. Без координат публикация отклоняется: по ним строятся гео-факты и досье, а нулевая точка — это точка в Гвинейском заливе.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListingID"
          }
        ],
        "responses": {
          "200": {
            "description": "Опубликовано.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            }
          },
          "401": {
            "description": "Ключ не передан, не распознан, отозван или истёк.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "invalid_api_key": {
                    "value": {
                      "error": {
                        "type": "authentication_error",
                        "code": "invalid_api_key",
                        "message": "Ключ API не распознан. Проверьте заголовок Authorization: Bearer hab_live_…",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ключу не хватает права, либо доступ партнёра приостановлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "insufficient_scope": {
                    "value": {
                      "error": {
                        "type": "permission_error",
                        "code": "insufficient_scope",
                        "message": "Ключу не хватает права search:read",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Превышен лимит. Заголовок Retry-After говорит, через сколько секунд повторять.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "type": "rate_limit_error",
                        "code": "rate_limit_exceeded",
                        "message": "Превышен лимит запросов (120 в минуту). Повторите через 12 с",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не прошёл проверку. Поле названо в param.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "validation_error": {
                    "value": {
                      "error": {
                        "type": "invalid_request_error",
                        "code": "validation_error",
                        "message": "query обязателен",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "query"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Объявление не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "owner_listing_not_found": {
                    "value": {
                      "error": {
                        "type": "invalid_request_error",
                        "code": "owner_listing_not_found",
                        "message": "Объявление не найдено",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/partner/v1/listings/{listing_id}/unpublish": {
      "post": {
        "operationId": "unpublishListing",
        "tags": [
          "Витрина"
        ],
        "summary": "Снять с публикации",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListingID"
          }
        ],
        "responses": {
          "200": {
            "description": "Снято.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            }
          },
          "401": {
            "description": "Ключ не передан, не распознан, отозван или истёк.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "invalid_api_key": {
                    "value": {
                      "error": {
                        "type": "authentication_error",
                        "code": "invalid_api_key",
                        "message": "Ключ API не распознан. Проверьте заголовок Authorization: Bearer hab_live_…",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ключу не хватает права, либо доступ партнёра приостановлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "insufficient_scope": {
                    "value": {
                      "error": {
                        "type": "permission_error",
                        "code": "insufficient_scope",
                        "message": "Ключу не хватает права search:read",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Превышен лимит. Заголовок Retry-After говорит, через сколько секунд повторять.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "type": "rate_limit_error",
                        "code": "rate_limit_exceeded",
                        "message": "Превышен лимит запросов (120 в минуту). Повторите через 12 с",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объявление не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "owner_listing_not_found": {
                    "value": {
                      "error": {
                        "type": "invalid_request_error",
                        "code": "owner_listing_not_found",
                        "message": "Объявление не найдено",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hab_live_…",
        "description": "Ключ передаётся заголовком `Authorization: Bearer hab_live_…`. Секрет показывается один раз при выпуске и в системе не хранится — восстановить его нельзя, только выпустить новый.\n\nПрава ключа задаются скоупами: `search:read`, `geo:read`, `listings:read`, `listings:write`, `leads:read`, `webhooks:write`."
      }
    },
    "parameters": {
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Курсор следующей страницы — берётся из next_cursor предыдущего ответа. Непрозрачен: собирать его самостоятельно не нужно."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        },
        "description": "Размер страницы. Больше 100 обрезается до 100."
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 255
        },
        "description": "Уникальная строка запроса. Повтор с тем же ключом вернёт первый ответ вместо второй операции; тот же ключ с другим телом — 409."
      },
...
    },
    "schemas": {
...
      "List": {
        "type": "object",
        "description": "Конверт списка. Пагинация курсорная: клиент листает, пока has_more не станет false, и никогда не собирает курсор сам.",
        "properties": {
          "object": {
            "type": "string",
            "description": "Всегда list.",
            "enum": [
              "list"
            ]
          },
          "item_object": {
            "type": "string",
            "description": "Тип элементов в data — то же значение, что в поле object каждого элемента."
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Элементы страницы."
          },
          "total": {
            "type": "integer",
            "description": "Сколько всего элементов в выборке, до пагинации."
          },
          "has_more": {
            "type": "boolean",
            "description": "Есть ли что-то за этой страницей."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Курсор следующей страницы. null, когда has_more = false."
          }
        },
        "required": [
          "object",
          "item_object",
          "data",
          "total",
          "has_more",
          "next_cursor"
        ]
      },
...
    }
  },
...
}